requirements.txt
Flask
gunicorn
#Python 2
#aiml
#Python 3
python-aiml